requirements.txt
pandas~=2.2.2
UnityPy~=1.20.18
pillow~=11.0.0
python-dateutil~=2.9.0.post0
opencv-python~=4.11.0.86
numpy~=2.1.1
matplotlib~=3.10.0